+++ layer/logged_data/utils.py
@@ -1,6 +1,6 @@
+import inspect
 from pathlib import Path
 from typing import Any, List, Optional
-import inspect
 
 
 def get_base_module_list(value: Any) -> List[str]:
